third_party/WebKit/LayoutTests/screen_orientation/orientationchange-event.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

var orientations = [
  'portrait-primary',
  'portrait-secondary',
  'landscape-primary',
  'landscape-secondary'
];

var currentIndex = orientations.indexOf(window.screen.orientation.type);
// Count the number of calls received from the EventHandler set on screen.orientation.onchange.
var orientationChangeEventHandlerCalls = 0;
// Count the number of calls received from the listener set with screen.orientation.addEventListene().
var orientationChangeEventListenerCalls = 0;

function getNextIndex() {
  return (currentIndex + 1) % orientations.length;
}

window.screen.orientation.onchange = function() {
  orientationChangeEventHandlerCalls++;
};

window.screen.orientation.addEventListener('change', function() {
  orientationChangeEventListenerCalls++;
});

for (var i = 0; i < 4; ++i) {
  test(function() {
    window.testRunner.setMockScreenOrientation(orientations[getNextIndex()]);

    currentIndex = getNextIndex();
    assert_equals(screen.orientation.type, orientations[currentIndex]);

    assert_equals(orientationChangeEventHandlerCalls, i + 1);
    assert_equals(orientationChangeEventListenerCalls, i + 1);
  }, "Test that orientationchange event is fired when the orientation changes");
}

test(function() {
  window.testRunner.setMockScreenOrientation(orientations[currentIndex]);
  assert_equals(screen.orientation.type, orientations[currentIndex]);

  assert_equals(orientationChangeEventHandlerCalls, 4);
  assert_equals(orientationChangeEventListenerCalls, 4);
}, "Test that orientationchange event is not fired when the orientation does not change");

</script>
</body>
</html>
